Add more default gems
The gems `base64` and `csv` is getting deprecation warnings
when i use the latest version of ruby. We use an older version
here so not really a problem but makes me more prepared for it. 

The gems `down`, `pdf-reader` and `roo` are commonly used
by us. Adding them will decrease the execution time when they
are needed.